playlists: add Create Playlist dialog item to AddToPlaylist menu
Fixes #145
