+++ src/components/AddToPlaylistMenu/index.js
@@ -1,107 +1,52 @@
-import find from 'array-find';
-import React, { Component, PropTypes } from 'react';
-import Paper from 'material-ui/Paper';
-import Menu from 'material-ui/Menu';
-import MenuItem from 'material-ui/MenuItem';
-// 😱
-import RenderToLayer from 'material-ui/internal/RenderToLayer';
+import * as React from 'react';
+import CreatePlaylistIcon from 'material-ui/svg-icons/content/add';
 
-const MENU_HEIGHT = 320;
-const MENU_WIDTH = 280;
+import PromptDialog from '../Dialogs/PromptDialog';
+import PlaylistsMenu from './PlaylistsMenu';
 
-const RANDOM_MUI_PADDING = 8;
-const SCROLLBAR_WIDTH = 7;
-
-const positionInsideWindow = (position, expectedHeight) => {
-  const constrained = { x: position.x, y: position.y };
-  const h = Math.min(expectedHeight, MENU_HEIGHT);
-  const w = MENU_WIDTH;
-  if (position.y + h >= window.innerHeight) {
-    // position at the bottom of the screen
-    constrained.y = window.innerHeight - h;
-  }
-  if (position.x + w >= window.innerWidth) {
-    // position to the left-hand side of the anchor, instead of the right-hand side
-    constrained.x -= w;
-  }
-  return constrained;
-};
-
-const menuStyle = {
-  textAlign: 'left',
-  zIndex: 30
-};
-const menuItemStyle = {
-  WebkitAppearance: 'initial'
-};
-
-export default class AddToPlaylistMenu extends Component {
+export default class AddToPlaylistMenu extends React.Component {
   static propTypes = {
-    onClose: PropTypes.func.isRequired,
-    onSelect: PropTypes.func.isRequired,
-    playlists: PropTypes.arrayOf(PropTypes.object),
-    position: PropTypes.shape({
-      x: PropTypes.number,
-      y: PropTypes.number
-    })
+    onClose: React.PropTypes.func.isRequired,
+    onSelect: React.PropTypes.func.isRequired,
+    onCreatePlaylist: React.PropTypes.func.isRequired
   };
 
-  handleSelect = (e, item) => {
-    const playlistID = item.props.value;
-    this.props.onClose();
-    this.props.onSelect(
-      find(this.props.playlists, pl => pl._id === playlistID)
-    );
+  state = {
+    creating: false
   };
 
-  renderLayer = () => {
-    const { playlists, position } = this.props;
-    const fixedPosition = positionInsideWindow(position, (playlists.length + 1) * 48);
-    return (
-      <div
-        style={{
-          position: 'absolute',
-          left: fixedPosition.x,
-          top: fixedPosition.y,
-          width: MENU_WIDTH + RANDOM_MUI_PADDING + SCROLLBAR_WIDTH
-        }}
-      >
-        <Paper>
-          <Menu
-            style={menuStyle}
-            maxHeight={MENU_HEIGHT}
-            width={MENU_WIDTH}
-            autoWidth={false}
-            onItemTouchTap={this.handleSelect}
-          >
-            <MenuItem
-              style={menuItemStyle}
-              primaryText="New Playlist"
-            />
-            {playlists.map(playlist => (
-              <MenuItem
-                key={playlist._id}
-                style={menuItemStyle}
-                value={playlist._id}
-                primaryText={playlist.name}
-                secondaryText={`${playlist.size || 0}`}
-                checked={!!playlist.active}
-              />
-            ))}
-          </Menu>
-        </Paper>
-      </div>
-    );
+  handleOpen = () => {
+    this.setState({ creating: true });
   };
+
+  handleClose = () => {
+    this.setState({ creating: false });
+  };
+
+  handleSubmit = playlistName =>
+    Promise.resolve(this.props.onCreatePlaylist(playlistName))
+      .then(playlist => this.props.onSelect(playlist))
+      .then(() => this.props.onClose());
 
   render() {
-    const { onClose } = this.props;
     return (
-      <RenderToLayer
-        open
-        componentClickAway={onClose}
-        render={this.renderLayer}
-      />
+      <div>
+        {!this.state.creating && (
+          <PlaylistsMenu
+            {...this.props}
+            onCreatePlaylist={this.handleOpen}
+          />
+        )}
+        {this.state.creating && (
+          <PromptDialog
+            title="Playlist Name"
+            icon={<CreatePlaylistIcon color="#777" />}
+            submitLabel="Create"
+            onSubmit={this.handleSubmit}
+            onCancel={this.handleClose}
+          />
+        )}
+      </div>
     );
   }
 }
